+++ src/main/simplejpa/artifact/repository/DefaultRepositoryClass.java
@@ -1,0 +1,12 @@
+package simplejpa.artifact.repository;
+
+import griffon.core.GriffonApplication;
+import org.codehaus.griffon.runtime.core.DefaultGriffonClass;
+
+public class DefaultRepositoryClass extends DefaultGriffonClass {
+
+    public DefaultRepositoryClass(GriffonApplication app, Class<?> clazz) {
+        super(app, clazz, "repository", "Repository");
+    }
+
+}