+++ src/templates/artifacts/Repository.groovy
@@ -1,0 +1,6 @@
+@artifact.package@import simplejpa.transaction.Transaction
+
+@Transaction
+class @artifact.name@ {
+
+}