+++ release_notes.md
@@ -156,3 +156,5 @@ First release into griffon artifacts portal.
 1. Allow user to add scaffolding configurations in `Config.groovy`.
 1. Add `griffon.simplejpa.scaffolding.auto` to automatically run scaffolding before compiling classes.
 1. Add DDD scaffolding generator that can be selected by using `generate-all -generator=simplejpa.scaffolding.generator.ddd.DDDGenerator`.
+1. Add new artifact type: repository.
+1. Add `create-repository` command.